#96: move hlc tests to one bundle
