+++ tests/org.eclipse.n4js.hlc.tests/META-INF/MANIFEST.MF
@@ -6,9 +6,11 @@ Bundle-Version: 1.0.0.qualifier
 Bundle-SymbolicName: org.eclipse.n4js.hlc.tests
 Export-Package: org.eclipse.n4js.hlc.tests
 Require-Bundle: org.eclipse.n4js,
- org.eclipse.n4js.hlc;bundle-version="1.0.0",
+ org.eclipse.n4js.hlc,
  org.eclipse.n4js.hlc.base,
  org.junit;bundle-version="4.11.0",
- com.google.guava;bundle-version="15.0.0"
+ com.google.guava;bundle-version="15.0.0",
+ org.eclipse.n4js.external.libraries,
+ org.eclipse.n4js.runner
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Import-Package: org.apache.log4j
