+++ tests/org.eclipse.n4js.hlc.tests/src/main/java/org/eclipse/n4js/hlc/tests/N4CliHelper.java
@@ -8,7 +8,7 @@
  * Contributors:
  *   NumberFour AG - Initial API and implementation
  */
-package org.eclipse.n4js.hlc.helper;
+package org.eclipse.n4js.hlc.tests;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.Assert.assertEquals;
