+++ tests/org.eclipse.n4js.hlc.tests/src/test/java/org/eclipse/n4js/hlc/tests/AT_IDEBUG_695_CannotSetFinalFieldInCtorForStaticPolyfillsTest.java
@@ -8,15 +8,14 @@
  * Contributors:
  *   NumberFour AG - Initial API and implementation
  */
-package org.eclipse.n4js.hlc;
+package org.eclipse.n4js.hlc.tests;
 
-import static org.eclipse.n4js.hlc.IncompleteApiImplementationTest.runCaptureOut;
+import static org.eclipse.n4js.hlc.tests.IncompleteApiImplementationTest.runCaptureOut;
 
 import java.io.IOException;
 
 import org.eclipse.n4js.hlc.base.N4jscBase;
 import org.eclipse.n4js.hlc.base.N4jscBase.ExitCodeException;
-import org.eclipse.n4js.hlc.helper.N4CliHelper;
 import org.junit.BeforeClass;
 import org.junit.Test;
