+++ tests/org.eclipse.n4js.hlc.tests/src/test/java/org/eclipse/n4js/hlc/tests/AbstractN4jscTest.java
@@ -10,111 +10,219 @@
  */
 package org.eclipse.n4js.hlc.tests;
 
-import static org.eclipse.n4js.hlc.tests.HlcTestingConstants.N4JSC_JAR;
-import static org.eclipse.n4js.hlc.tests.HlcTestingConstants.TARGET;
+import static com.google.common.base.Preconditions.checkState;
+import static com.google.common.collect.Sets.newHashSet;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Collections;
+import java.nio.file.FileVisitResult;
+import java.nio.file.FileVisitor;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.attribute.BasicFileAttributes;
+import java.util.Collection;
+import java.util.concurrent.atomic.AtomicInteger;
 
-import org.junit.After;
-import org.junit.Before;
-
-import org.eclipse.n4js.hlc.helper.N4CliHelper;
+import org.eclipse.n4js.N4JSGlobals;
+import org.eclipse.n4js.utils.collections.Arrays2;
 import org.eclipse.n4js.utils.io.FileCopier;
 import org.eclipse.n4js.utils.io.FileDeleter;
+import org.eclipse.n4js.validation.helper.N4JSLanguageConstants;
+import org.junit.After;
+
+import com.google.common.base.Predicate;
+import com.google.common.base.Predicates;
 
 /**
  */
 public abstract class AbstractN4jscTest {
 
-	// Running directory will be ${TARGET}/${WSP}
-
-	/** Sub folder in target folder. */
+	/**
+	 * name of sub-directory of the compiled result. Used to count files having this string as directory name on it's
+	 * {@link #assertFilesCompiledToES(int, String)}
+	 */
+	private static final String SUBGENERATOR_PATH = N4JSLanguageConstants.TRANSPILER_SUBFOLDER_FOR_TESTS;
+	/** name of target folder */
+	protected static final String TARGET = "target";
+	/** name of workspace sub-folder (inside target folder) */
 	protected static final String WSP = "wsp";
-
-	/** source of test data, will be copied to TARGET/WSP */
-	protected final String fixture;
-
-	/**
-	 * Output-log-file of external Process. The file will be assigned given on the current test-method by calling
-	 * {@link #logFile()} as first statement in each test-method.
-	 */
-	protected File outputLogFile = null;
+	/** name of package containing the test resources */
+	protected static final String FIXTURE = "src/test/resources";
+	/** name of default test data set */
+	protected static final String TEST_DATA_SET__BASIC = "basic";
+	/** name of test data set for launching testers from the command line */
+	protected static final String TEST_DATA_SET__TESTERS = "testers";
 
 	/**
-	 * Subclass must provide the fixture, i.e. name of folder containing test data.
+	 * Copy a fresh fixture to the workspace area. Deleting old leftovers from former tests.
+	 *
+	 * @returns file indicating the relative path to the copied data set
 	 */
-	protected AbstractN4jscTest(String fixture) {
-		this.fixture = fixture;
+	protected static File setupWorkspace(String testDataSet) throws IOException {
+		return setupWorkspace(testDataSet, Predicates.alwaysFalse());
 	}
 
 	/**
-	 * Copy a fresh fixture to the workspace area.
+	 * Copy a fresh fixture to the workspace area. Deleting old leftovers from former tests.
+	 *
+	 * @returns file indicating the relative path to the copied data set
 	 */
-	@Before
-	public void setupWorkspace() throws IOException {
+	protected static File setupWorkspace(String testDataRoot, String testDataSet) throws IOException {
+		return setupWorkspace(testDataRoot, testDataSet, Predicates.alwaysFalse());
+	}
+
+	/**
+	 * Copy a fresh fixture to the workspace area. Deleting old leftovers from former tests. Also includes all N4JS
+	 * libraries from the {@code n4js} Git repository which name provides {@code true} value for the given predicate.
+	 *
+	 * @returns file indicating the relative path to the copied data set
+	 */
+	protected static File setupWorkspace(String testDataSet,
+			Predicate<String> n4jsLibrariesPredicate)
+			throws IOException {
+		return setupWorkspace(FIXTURE, testDataSet, n4jsLibrariesPredicate);
+	}
+
+	/**
+	 * Copy a fresh fixture to the workspace area. Deleting old leftovers from former tests. Also includes all N4JS
+	 * libraries from the {@code n4js} Git repository which name provides {@code true} value for the given predicate.
+	 *
+	 * @returns file indicating the relative path to the copied data set
+	 */
+	protected static File setupWorkspace(String testDataRoot, String testDataSet,
+			Predicate<String> n4jsLibrariesPredicate)
+			throws IOException {
 		File wsp = new File(TARGET, WSP);
-		File fixtureFile = new File(fixture);
-
-		System.out.println("BEFORE: 			  current root " + new File(".").getAbsolutePath());
-		System.out.println("BEFORE: current workspace would be " + wsp.getAbsolutePath());
-
+		File fixture = new File(testDataRoot, testDataSet);
 		// clean
+		// System.out.println("Workspace : " + wsp.getAbsolutePath());
 		// Files.deleteIfExists(wsp.toPath());
-		FileDeleter.delete(wsp.toPath());
-		// copy
-		FileCopier.copy(fixtureFile.toPath(), wsp.toPath());
+		if (wsp.exists()) {
+			FileDeleter.delete(wsp.toPath(), true);
+		}
 
+		// copy fixtures to workspace
+		FileCopier.copy(fixture.toPath(), wsp.toPath(), true);
+
+		final File gitRoot = new File(new File("").getAbsolutePath()).getParentFile().getParentFile();
+		final File n4jsLibraryRoot = new File(gitRoot, N4JSGlobals.SHIPPED_CODE_SOURCES_FOLDER_NAME);
+		final File[] n4jsLibraries = n4jsLibraryRoot.listFiles();
+		// copy N4JS libraries on demand
+		if (!Arrays2.isEmpty(n4jsLibraries)) {
+			for (final File n4jsLibrary : n4jsLibraries) {
+				if (n4jsLibrariesPredicate.apply(n4jsLibrary.getName())) {
+					System.out.println("Including N4JS library in workspace: '" + n4jsLibrary.getName() + "'.");
+					final File libFolder = new File(wsp, n4jsLibrary.getName());
+					libFolder.mkdir();
+					checkState(libFolder.isDirectory(),
+							"Error while copying N4JS library '" + n4jsLibrary.getName() + "' to workspace.");
+					FileCopier.copy(n4jsLibrary.toPath(), libFolder.toPath(), true);
+				}
+			}
+		}
+
+		return wsp;
 	}
 
 	/**
-	 * Append external process output to the junit std-out content.
+	 * Flush outputs after tests.
 	 */
 	@After
-	public void appendExternalOutpouToStdout() {
-		N4CliHelper.appendExternalOutputToStdout(outputLogFile);
+	public void flush() {
+		System.out.flush();
+		System.err.flush();
+		System.out.println("======= ======= ======= ======= ======= ======= ======= ======= ======= ======= ");
+		System.out.flush();
+
 	}
 
 	/**
-	 * Deletes the project with the given name (its folder and all contained files and sub folders) from the temporary
-	 * workspace directory. This can be used to change the test data by removing selected projects at the beginning of a
-	 * test method.
+	 * @param expectedCompiledModuleCount
+	 *            expected number of compiled '.js' files in the {@value #SUBGENERATOR_PATH} folder.
+	 * @param workspaceRootPath
+	 *            subtree to search in passed as argument to {@link File}
 	 */
-	protected void deleteProject(String projectId) throws IOException {
-		File wsp = new File(TARGET, WSP);
-		File project = new File(wsp, projectId);
-		FileDeleter.delete(project.toPath());
+	protected void assertFilesCompiledToES(int expectedCompiledModuleCount, String workspaceRootPath) {
+		assertEquals("Files expected", expectedCompiledModuleCount, countFilesCompiledToES(workspaceRootPath));
 	}
 
 	/**
-	 * Create & start java-Process calling n4jsc.jar jar with args from {@value #TARGET}-folder.
+	 * Counts the number of files ending in .js in the {@value #SUBGENERATOR_PATH} folder.
 	 *
-	 * @param args
-	 *            arguments to pass after jar - call
-	 *
-	 * @return running process
-	 * @throws IOException
-	 *             if errored.
+	 * @param workspaceRootPath
+	 *            the directory to recursively search
+	 * @return the number of files ending in .js
 	 */
-	protected Process createAndStartProcess(String... args) throws IOException {
-		ArrayList<String> args2 = new ArrayList<>();
-		// Collections.addAll(args2, "java", "-jar", TARGET + "/" + N4JSC_JAR);
-		Collections.addAll(args2, "java", "-jar", N4JSC_JAR);
-		Collections.addAll(args2, args);
-		return N4CliHelper.createAndStartProcessIntern(outputLogFile, TARGET, args2.toArray(new String[args2.size()]));
+	protected int countFilesCompiledToES(String workspaceRootPath) {
+		final File workspaceRoot = new File(workspaceRootPath);
+
+		final File gitRoot = new File(new File("").getAbsolutePath()).getParentFile().getParentFile();
+		final File n4jsLibrariesRoot = new File(gitRoot, N4JSGlobals.SHIPPED_CODE_SOURCES_FOLDER_NAME);
+		final Collection<String> n4jsLibraryNames = newHashSet(n4jsLibrariesRoot.list());
+
+		final AtomicInteger counter = new AtomicInteger();
+		try {
+			Files.walkFileTree(workspaceRoot.toPath(), new FileVisitor<Path>() {
+
+				@Override
+				public FileVisitResult preVisitDirectory(Path dir, BasicFileAttributes attrs) throws IOException {
+
+					// skip src
+					if ("src".equals(dir.getFileName())) {
+						return FileVisitResult.SKIP_SUBTREE;
+					}
+
+					if (n4jsLibraryNames.contains(dir.toFile().getName())) {
+						return FileVisitResult.SKIP_SUBTREE;
+					}
+
+					return FileVisitResult.CONTINUE;
+				}
+
+				@Override
+				public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
+					if (file.getFileName().toString().endsWith(".js")) {
+						for (int j = 0; j < file.getNameCount() - 1; j++) {
+							if (SUBGENERATOR_PATH.equals(file.getName(j).toString())) {
+								counter.incrementAndGet();
+								return FileVisitResult.CONTINUE;
+							}
+						}
+					}
+					return FileVisitResult.CONTINUE;
+				}
+
+				@Override
+				public FileVisitResult visitFileFailed(Path file, IOException exc) throws IOException {
+					return FileVisitResult.CONTINUE;
+				}
+
+				@Override
+				public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException {
+					if (SUBGENERATOR_PATH.equals(dir.getFileName()))
+						return FileVisitResult.SKIP_SIBLINGS;
+					return FileVisitResult.CONTINUE;
+				}
+			});
+		} catch (IOException e) {
+			e.printStackTrace();
+			assertTrue("Could not parse file tree of path '" + workspaceRootPath + "' exc = " + e.getMessage(), false);
+		}
+
+		return counter.get();
 	}
 
 	/**
-	 * Should be called as first line in test-mehtods.
+	 * Determine class.method of the direct caller by inspecting the current thread's stack trace.
 	 *
-	 * Creates an log-file in the {@link #TARGET}-folder based on the callers Class/Methodname. something like
-	 * "target/org.eclipse.n4js.hlc.test.N4jscSingleFileCompileIT.testHelp.log"
+	 * @return Class + Method name of the caller.
 	 */
-	protected void logFile() {
+	protected String logMethodname() {
 		StackTraceElement stackTraceElement = Thread.currentThread().getStackTrace()[2];
-		String name = stackTraceElement.getClassName() + "." + stackTraceElement.getMethodName() + ".log";
-		outputLogFile = new File(TARGET, name);
+		String name = stackTraceElement.getClassName() + "." + stackTraceElement.getMethodName() + "";
+		return "INSIDE of " + name;
 	}
+
 }
