+++ tests/org.eclipse.n4js.hlc.tests/src/test/java/org/eclipse/n4js/hlc/tests/InstallCompileRunN4jscExternalTest.java
@@ -8,10 +8,10 @@
  * Contributors:
  *   NumberFour AG - Initial API and implementation
  */
-package org.eclipse.n4js.hlc;
+package org.eclipse.n4js.hlc.tests;
 
 import static java.util.Collections.singletonMap;
-import static org.eclipse.n4js.hlc.IncompleteApiImplementationTest.runCaptureOut;
+import static org.eclipse.n4js.hlc.tests.IncompleteApiImplementationTest.runCaptureOut;
 import static org.eclipse.n4js.runner.SystemLoaderInfo.COMMON_JS;
 
 import java.io.IOException;
@@ -19,13 +19,12 @@ import java.util.Map;
 
 import org.eclipse.n4js.hlc.base.N4jscBase.BuildType;
 import org.eclipse.n4js.hlc.base.N4jscBase.ExitCodeException;
-import org.eclipse.n4js.hlc.helper.N4CliHelper;
 import org.junit.Test;
 
 /**
  * Downloads, installs, compiles and runs 'express'.
  */
-public class InstallCompileRunN4jscExternalWithSingleFileCompileTest extends BaseN4jscExternalTest {
+public class InstallCompileRunN4jscExternalTest extends BaseN4jscExternalTest {
 
 	@Override
 	protected Map<String, String> getNpmDependencies() {
@@ -39,10 +38,9 @@ public class InstallCompileRunN4jscExternalWithSingleFileCompileTest extends Bas
 	@Test
 	public void testCompileAndRunWithExternalDependencies() throws IOException, ExitCodeException {
 		System.out.println(name.getMethodName());
-		setupWorkspace("external_singleProjectOrFileCompile");
+		setupWorkspace("external");
 		final String wsRoot = TARGET + "/" + WSP;
-		final String fileToCompile = wsRoot + "/external.project/src/Main.n4js";
-		final String fileToRun = fileToCompile;
+		final String fileToRun = wsRoot + "/external.project/src/Main.n4js";
 
 		final String[] args = {
 				"--systemLoader", COMMON_JS.getId(),
@@ -52,12 +50,12 @@ public class InstallCompileRunN4jscExternalWithSingleFileCompileTest extends Bas
 				"-r", fileToRun,
 				"--debug",
 				"--verbose",
-				"-t", BuildType.singlefile.toString(),
-				fileToCompile
+				"--projectlocations", wsRoot,
+				"-t", BuildType.allprojects.toString()
 		};
 		final String out = runCaptureOut(args);
 		N4CliHelper.assertExpectedOutput(
-				"Application was created!", out);
+				"express properties: application, request, response, Route, Router, query, static", out);
 	}
 
 }
