+++ tests/org.eclipse.n4js.hlc.tests/src/test/java/org/eclipse/n4js/hlc/tests/N4jscTestersTest.java
@@ -8,7 +8,7 @@
  * Contributors:
  *   NumberFour AG - Initial API and implementation
  */
-package org.eclipse.n4js.hlc;
+package org.eclipse.n4js.hlc.tests;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.fail;
