+++ tests/org.eclipse.n4js.hlc.tests/src/test/java/org/eclipse/n4js/hlc/tests/TargetPlatformRangeConfigurationsTest.java
@@ -8,10 +8,10 @@
  * Contributors:
  *   NumberFour AG - Initial API and implementation
  */
-package org.eclipse.n4js.hlc;
+package org.eclipse.n4js.hlc.tests;
 
 import static java.util.Collections.singletonMap;
-import static org.eclipse.n4js.hlc.IncompleteApiImplementationTest.runCaptureOut;
+import static org.eclipse.n4js.hlc.tests.IncompleteApiImplementationTest.runCaptureOut;
 import static org.eclipse.n4js.runner.SystemLoaderInfo.COMMON_JS;
 
 import java.io.IOException;
@@ -19,30 +19,27 @@ import java.util.Map;
 
 import org.eclipse.n4js.hlc.base.N4jscBase.BuildType;
 import org.eclipse.n4js.hlc.base.N4jscBase.ExitCodeException;
-import org.eclipse.n4js.hlc.helper.N4CliHelper;
 import org.junit.Test;
 
 /**
- * Downloads, installs, compiles and runs 'express'.
+ * Downloads, installs, compiles and runs 'express' where target platform file specifies version range.
  */
-public class InstallCompileRunN4jscExternalWithSingleProjectCompileTest extends BaseN4jscExternalTest {
+public class TargetPlatformRangeConfigurationsTest extends BaseN4jscExternalTest {
 
 	@Override
 	protected Map<String, String> getNpmDependencies() {
-		return singletonMap("express", "@4.13.4");
+		return singletonMap("express", "@\">=4.0.0 <5.0.0\"");
 	}
 
 	/**
-	 * Test for checking the npm support in the headless case by downloading third party package, importing it and
-	 * running it with Common JS.
+	 * Test for compiling project with external dependency specified within version range.
 	 */
 	@Test
 	public void testCompileAndRunWithExternalDependencies() throws IOException, ExitCodeException {
 		System.out.println(name.getMethodName());
-		setupWorkspace("external_singleProjectOrFileCompile");
+		setupWorkspace("external");
 		final String wsRoot = TARGET + "/" + WSP;
-		final String projectToCompile = wsRoot + "/external.project";
-		final String fileToRun = projectToCompile + "/src/Main.n4js";
+		final String fileToRun = wsRoot + "/external.project/src/Main.n4js";
 
 		final String[] args = {
 				"--systemLoader", COMMON_JS.getId(),
@@ -52,12 +49,11 @@ public class InstallCompileRunN4jscExternalWithSingleProjectCompileTest extends 
 				"-r", fileToRun,
 				"--debug",
 				"--verbose",
-				"-t", BuildType.projects.toString(),
-				projectToCompile
+				"--projectlocations", wsRoot,
+				"-t", BuildType.allprojects.toString()
 		};
 		final String out = runCaptureOut(args);
 		N4CliHelper.assertExpectedOutput(
-				"Application was created!", out);
+				"express properties: application, request, response, Route, Router, query, static", out);
 	}
-
 }
