+++ tools/org.eclipse.n4js.hlc/META-INF/MANIFEST.MF
@@ -4,8 +4,7 @@ Bundle-Name: %pluginName
 Bundle-Vendor: %providerName
 Bundle-Version: 1.0.0.qualifier
 Bundle-SymbolicName: org.eclipse.n4js.hlc
-Export-Package: org.eclipse.n4js.hlc,
- org.eclipse.n4js.hlc.helper
+Export-Package: org.eclipse.n4js.hlc
 Require-Bundle: org.eclipse.n4js,
  org.eclipse.n4js.hlc.base,
  org.eclipse.n4js.n4mf,
